MicroROS and the firmware now build successfully.

+++ Source Code/pico_ws/libmicroros/build_uros.sh
@@ -10,25 +10,44 @@ echo "MicroROS Library Build Script."
 echo "NOTE: DO NOT RUN AS ROOT, RUN AS 'urosdev'!"
 echo ""
 
-read -p "This will delete the previous build and start from scratch. Continue? (y/n) " confirm
+while getopts bf flag
+do
+    case "${flag}" in
+        b) FULL_REBUILD="false";;
+        f) FULL_REBUILD="true";;
+        *) echo "Invalid flags! (-b: firmware build only, -f: full re-build)" && exit 1;;
+    esac
+done
 
-case $confirm in 
-    [yY]) 
-        cd ~/pico_ws/uros_ws \
-        && echo "Removing directories..." && rm -rf ./build && rm -rf ./firmware && rm -rf ./install && rm -rf ./log \
-        && echo "Building..." && colcon build --packages-select rrp_pico_coms \
-        && source install/local_setup.bash \
-        && sudo apt update \
-        && ros2 run micro_ros_setup create_firmware_ws.sh generate_lib \
-        && ros2 run rrp_pico_coms create_fwws.sh \
-        && ros2 run micro_ros_setup build_firmware.sh "$(pwd)/my_toolchain.cmake" "$(pwd)/my_colcon.meta"
-        echo "Build completed successfully (probably)!"
-        ;;
-    [nN]) 
-        echo "Cancelled."
-        ;;
-    *)
-        echo "Invalid input!"
-        echo "Cancelled."
-        ;;
-esac
+if [ "$FULL_REBUILD" == "true" ]; then
+    read -p "This will delete the previous build and start from scratch. Continue? (y/n) " confirm
+
+    case $confirm in 
+        [yY]) 
+            cd ~/pico_ws/uros_ws \
+            && echo "Removing directories..." && rm -rf ./build && rm -rf ./firmware && rm -rf ./install && rm -rf ./log \
+            && echo "Building..." && colcon build --packages-select rrp_pico_coms \
+            && source install/local_setup.bash \
+            && sudo apt update \
+            && ros2 run micro_ros_setup create_firmware_ws.sh generate_lib \
+            && ros2 run rrp_pico_coms create_fwws.sh \
+            && ros2 run micro_ros_setup build_firmware.sh "$(pwd)/my_toolchain.cmake" "$(pwd)/my_colcon.meta"
+            echo "Build completed successfully (probably)!"
+            ;;
+        [nN]) 
+            echo "Cancelled."
+            ;;
+        *)
+            echo "Invalid input!"
+            echo "Cancelled."
+            exit 1
+            ;;
+    esac
+fi
+
+if [ "$FULL_REBUILD" == "false" ]; then
+    echo "Re-building firmware only..."
+    cd ~/pico_ws/uros_ws \
+    && ros2 run micro_ros_setup build_firmware.sh "$(pwd)/my_toolchain.cmake" "$(pwd)/my_colcon.meta"
+    echo "Build completed successfully (probably)!"
+fi
